+++ .test/auth.tf
@@ -5,6 +5,7 @@ provider "azurerm" {
       prevent_deletion_if_contains_resources = false
     }
   }
+  storage_use_azuread = true
 }
 
 # Versions
@@ -14,7 +15,7 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "3.74.0"
+      version = "4.10.0"
     }
   }