+++ .test/test.tf
@@ -54,8 +54,8 @@ module "azurerm-network-vnet" {
     default2 = {
       name                                          = "snet-tde3-ic-default2"
       address_prefixes                              = ["10.0.1.0/24"]
-      private_endpoint_network_policies_enabled     = true
-      private_link_service_network_policies_enabled = true
+      # private_endpoint_network_policies_enabled     = true
+      # private_link_service_network_policies_enabled = true
     }
   }